# daemon/codereview/router.py
"""Daemon code review router — participant + host endpoints."""
import asyncio
import json
import logging
import os
from typing import Optional

from fastapi import APIRouter, Request
from fastapi.responses import JSONResponse
from pydantic import BaseModel

from daemon.codereview.state import codereview_state
from daemon.participant.state import participant_state
from daemon.scores import scores
from daemon.ws_publish import broadcast, broadcast_event, host_event
from daemon.ws_messages import (
    ActivityUpdatedMsg,
    CodereviewOpenedMsg,
    CodereviewSelectionClosedMsg,
    CodereviewSelectionsUpdatedMsg,
    CodereviewLineConfirmedMsg,
    CodereviewClearedMsg,
    ScoresUpdatedMsg,
)

logger = logging.getLogger(__name__)


class OkResponse(BaseModel):
    ok: bool = True


class SelectionRequest(BaseModel):
    lines: list[int] = []


class CreateCodeReviewRequest(BaseModel):
    snippet: str
    language: Optional[str] = None
    smart_paste: bool = True


class SetCodeReviewStatusRequest(BaseModel):
    open: bool = True


class ConfirmLineRequest(BaseModel):
    line: int


class OkWithPhaseResponse(BaseModel):
    ok: bool = True
    phase: str


class ConfirmLineResponse(BaseModel):
    ok: bool = True
    confirmed_line: int


# ── Smart paste logic (copied from features/codereview/router.py) ──

_EXTRACT_PROMPT = """Extract only the code snippet from the following text.
Remove any markdown formatting, explanations, comments about the code, or surrounding text.
Return ONLY a JSON object with two fields:
- "code": the extracted code (preserve original indentation, no markdown fences)
- "language": the programming language as lowercase identifier (one of: java, python, javascript, typescript, sql, go, csharp, kotlin, bash, or null if unknown)

If the input is already clean code with no surrounding text, return it as-is in the JSON format."""

_SMART_PASTE_INPUT_LIMIT = 10000


def _extract_code_with_ai(raw_snippet: str) -> tuple[str, str | None] | None:
    """Call Claude Haiku to extract code from LLM output.
    Returns (code, language) or None on any failure."""
    try:
        from daemon.llm.adapter import create_message
        api_key = os.environ.get("ANTHROPIC_API_KEY")
        if not api_key:
            return None

        truncated = raw_snippet[:_SMART_PASTE_INPUT_LIMIT]
        response = create_message(
            api_key=api_key,
            model="claude-haiku-4-5-20251001",
            max_tokens=4096,
            messages=[{"role": "user", "content": f"{_EXTRACT_PROMPT}\n\n---\n\n{truncated}"}],
            timeout=5.0,
        )

        text = response.content[0].text.strip()
        # Strip markdown fences if Claude wrapped the JSON
        if text.startswith("```"):
            text = text.split("\n", 1)[1] if "\n" in text else text[3:]
        if text.endswith("```"):
            text = text[:text.rfind("```")]
        text = text.strip()

        result = json.loads(text)
        code = result.get("code", "").strip()
        if not code:
            return None
        language = result.get("language")
        if isinstance(language, str):
            language = language.lower()
        logger.info("Smart paste extracted %d lines, language=%s", len(code.splitlines()), language)
        return (code, language)
    except Exception:
        logger.debug("Smart paste extraction failed", exc_info=True)
        return None


# ── Participant router (proxied via Railway) ──

participant_router = APIRouter(prefix="/api/participant/codereview", tags=["codereview"])


@participant_router.put("/selection", response_model=OkResponse)
async def update_selection(request: Request, body: SelectionRequest):
    """Participant sets their selected lines (full replacement)."""
    pid = request.headers.get("x-participant-id")
    if not pid:
        return JSONResponse({"error": "Missing X-Participant-ID"}, status_code=400)

    # Activity gate
    if participant_state.current_activity != "codereview":
        return JSONResponse({"error": "Code review not active"}, status_code=409)

    # Phase gate
    if codereview_state.phase != "selecting":
        return JSONResponse({"error": "Selection phase is closed"}, status_code=409)

    codereview_state.select_lines(pid, body.lines)

    # Send selection update only to host (line counts are host-only)
    line_counts = {
        str(ln): sum(1 for sel in codereview_state.selections.values() if ln in sel)
        for pid_sel in codereview_state.selections.values()
        for ln in pid_sel
    }
    request.state.write_back_events = [
        host_event(CodereviewSelectionsUpdatedMsg(line_counts=line_counts)),
    ]

    return OkResponse()


# ── Host router (called directly on daemon localhost) ──
# NOTE: Host JS calls API('/codereview') which expands to /api/{session_id}/codereview.
# The prefix includes {session_id} path parameter to match this pattern.

host_router = APIRouter(prefix="/api/{session_id}/host/codereview", tags=["codereview"])

_MAX_LINES = 50


@host_router.post("", response_model=OkResponse)
async def create_codereview(body: CreateCodeReviewRequest):
    """Host creates a code review session."""
    snippet = body.snippet.strip()
    language = body.language  # None means auto-detect
    smart_paste = body.smart_paste

    if not snippet:
        return JSONResponse({"error": "Snippet cannot be empty"}, status_code=400)

    detected_language = None
    if smart_paste:
        result = await asyncio.to_thread(_extract_code_with_ai, snippet)
        if result:
            snippet, detected_language = result

    lines = snippet.splitlines()
    if len(lines) > _MAX_LINES:
        return JSONResponse({"error": f"Snippet cannot exceed {_MAX_LINES} lines"}, status_code=400)

    # Use detected language only if host chose "Auto-detect" (null)
    final_language = language if language is not None else detected_language

    participant_state.current_activity = "codereview"
    codereview_state.create(snippet, final_language)

    broadcast(CodereviewOpenedMsg(snippet=snippet, language=final_language))
    broadcast(ActivityUpdatedMsg(current_activity="codereview"))

    return OkResponse()


@host_router.put("/status", response_model=OkWithPhaseResponse)
async def set_codereview_status(body: SetCodeReviewStatusRequest):
    """Host closes the selection phase."""
    if not body.open:
        codereview_state.close_selection()
        broadcast(CodereviewSelectionClosedMsg())

    return OkWithPhaseResponse(phase=codereview_state.phase)


@host_router.put("/confirm-line", response_model=ConfirmLineResponse)
async def confirm_line(body: ConfirmLineRequest):
    """Host confirms a line as problematic and awards points."""
    if not codereview_state.snippet:
        return JSONResponse({"error": "No code review created yet"}, status_code=400)

    line_count = len(codereview_state.snippet.splitlines())
    if body.line < 0 or body.line >= line_count:
        return JSONResponse({"error": f"Line {body.line} is out of range (0–{line_count - 1})"}, status_code=400)

    if body.line in codereview_state.confirmed:
        return JSONResponse({"error": f"Line {body.line} is already confirmed"}, status_code=409)

    awarded_pids = codereview_state.confirm_line(body.line)

    for pid in awarded_pids:
        scores.add_score(pid, 200)

    broadcast(CodereviewLineConfirmedMsg(line=body.line))
    broadcast(ScoresUpdatedMsg(scores=scores.snapshot()))

    return ConfirmLineResponse(confirmed_line=body.line)


@host_router.delete("", response_model=OkResponse)
async def clear_codereview():
    """Host clears the code review."""
    codereview_state.clear()
    participant_state.current_activity = "none"

    broadcast(CodereviewClearedMsg())
    broadcast(ActivityUpdatedMsg(current_activity="none"))

    return OkResponse()