# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Daemon Host Panel
  version: 0.1.0
paths:
...
  /api/participant/codereview/selection:
    put:
      tags:
      - codereview
      summary: Update Selection
      description: Participant sets their selected lines (full replacement).
      operationId: update_selection_api_participant_codereview_selection_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: codereview
  /api/{session_id}/host/codereview:
    post:
      tags:
      - codereview
      summary: Create Codereview
      description: Host creates a code review session.
      operationId: create_codereview_api__session_id__host_codereview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCodeReviewRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: codereview
    delete:
      tags:
      - codereview
      summary: Clear Codereview
      description: Host clears the code review.
      operationId: clear_codereview_api__session_id__host_codereview_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: codereview
  /api/{session_id}/host/codereview/status:
    put:
      tags:
      - codereview
      summary: Set Codereview Status
      description: Host closes the selection phase.
      operationId: set_codereview_status_api__session_id__host_codereview_status_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCodeReviewStatusRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkWithPhaseResponse'
      x-feature: codereview
  /api/{session_id}/host/codereview/confirm-line:
    put:
      tags:
      - codereview
      summary: Confirm Line
      description: Host confirms a line as problematic and awards points.
      operationId: confirm_line_api__session_id__host_codereview_confirm_line_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmLineRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmLineResponse'
      x-feature: codereview
  /api/{session_id}/host/activity:
    put:
      tags:
      - activity
      summary: Set Activity
      description: Host switches the current activity.
      operationId: set_activity_api__session_id__host_activity_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetActivityRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetActivityResponse'
      x-feature: activity
    post:
      tags:
      - activity
      summary: Set Activity
      description: Host switches the current activity.
      operationId: set_activity_api__session_id__host_activity_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetActivityRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetActivityResponse'
      x-feature: activity
...
components:
  schemas:
...
    ConfirmLineRequest:
      properties:
        line:
          type: integer
          title: Line
      type: object
      required:
      - line
      title: ConfirmLineRequest
    ConfirmLineResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        confirmed_line:
          type: integer
          title: Confirmed Line
      type: object
      required:
      - confirmed_line
      title: ConfirmLineResponse
    CreateCodeReviewRequest:
      properties:
        snippet:
          type: string
          title: Snippet
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        smart_paste:
          type: boolean
          title: Smart Paste
          default: true
      type: object
      required:
      - snippet
      title: CreateCodeReviewRequest
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
    OkWithPhaseResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        phase:
          type: string
          title: Phase
      type: object
      required:
      - phase
      title: OkWithPhaseResponse
...
    SelectionRequest:
      properties:
        lines:
          items:
            type: integer
          type: array
          title: Lines
          default: []
      type: object
      title: SelectionRequest
...
    SetActivityRequest:
      properties:
        activity:
          type: string
          title: Activity
      type: object
      required:
      - activity
      title: SetActivityRequest
    SetActivityResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        current_activity:
          type: string
          title: Current Activity
      type: object
      required:
      - current_activity
      title: SetActivityResponse
    SetCodeReviewStatusRequest:
      properties:
        open:
          type: boolean
          title: Open
          default: true
      type: object
      title: SetCodeReviewStatusRequest
...